fix(platform): full remediation pass — billing/legal correctness, live-path data loss, honest security posture, marketing purge, ops wiring
Billing/legal (EUR everywhere; Wise; no placeholders on legal documents):
- invoice PDF: real legal identity (Bel Consulting OÜ, Sakala 7-2, EE102951727),
  seller/bank/buyer from data; bank=Wise via config, IBAN/BIC omitted when unset;
  negative-amount money formatter fixed; fonts aligned to shipped Noto Sans
- DEFAULT_BILLING_CURRENCY EUR; all \$ formatters → €; integer-cents formatting everywhere
- reverse-charge note only when reverse charge actually applied (valid VAT + 0% + non-EE)
- single VAT allocator (round half-up, reconcile final line) shared by both invoice writers;
  admin invoices use the sequential platform sequence
- fractional statutory VAT rates (FI 25.5%); invoices.vat_rate → double precision (migration 122);
  VAT summaries gate on VAT-number validity; dead estonia_vat_rate config removed
- proration preview: floor elapsed, integer math via proration::prorated_amount, € messages
- Stripe: sk_test_* rejected in production paths; checkout/portal idempotency keys;
  Stripe invoice.paid stores real line items
- OVERAGE IMPLEMENTED (was promised, never built): paid plans send into an overage
  allowance (OVERAGE_ALLOWANCE_PERCENT, default +100%) instead of a hard 403; a daily
  idempotent sweep invoices accrued overage at €0.40/1,000 per ended billing cycle
- pricing mirrors corrected (tools/lib/pricing.py EUR + real limits); training corpus
  swept to EUR with canonical limits; validators extended to cover every artifact

Live-path correctness:
- SMTP submission preserves attachments and custom headers through email_queue
  (worker contract: headers JSONB + attachments JSONB; regression test)
- generic 5xx no longer suppresses recipients tenant-wide (address-invalid proof required)
- worker supervised: processors restart with backoff instead of dying at startup
- circuit-breaker requeues its claim instead of stranding rows for the lease
- bounce attribution uses typed UUID comparison (no partitioned-table seq scan)
- submission EHLO rejects invalid args with 501 5.5.4 (parity with inbound)

Security honesty (was: 45k LOC of unwired "protection"):
- broken live_analytics test fixed; 12 unwired-crate READMEs banner NOT-WIRED
- ato-protection redis-lockout enabled by default; degraded state is a construction error
- login/signup enumeration closed (status/SSO policy post-verification; generic signup reply)
- SSO auto-provision uses VARCHAR(26) ids (signup 500 fixed); console api-key/dedicated-IP
  INSERTs match schema; console keys use keyed HMAC (parity with JSON path)
- CP gate appends (not replaces) Set-Cookie — admin flash messages reach the browser again
- password change revokes sessions; sales-autopilot refuses production start without a
  tenant allowlist; parse_duration multibyte panic; API-key modulo bias; bulk_create chunking

Ops (dead rules and missing services now real):
- analytics-worker DEPLOYED (Dockerfile stage, compose, CI/deploy/verify lists) — Postgres
  events retention/compaction actually runs
- pdf-renderer DEPLOYED to production (was dev-only while billing pointed at it — invoice
  PDFs silently failed); built from the mail-server context
- worker exports apexmail_email_queue_depth{status} — the backlog alerts were dead rules;
  DLQ alerts added; infrastructure-alerts false-coverage note corrected
- STARTTLS cert-expiry probes for SMTP/IMAP ports; stop_grace_period 90s on mta/imap;
  redis-exporter prod auth via JSON-map Docker secret (dev env documented);
  loopback-binding inversion fixed; migration-checksum rewrite requires typed confirmation;
  alert-receiver envs documented; DEPLOYMENT.md deploy model corrected (no GH Actions)

Marketing/site (internal boundary + consistency):
- GDPR "Confirm active deployment" placeholders replaced with real transfer safeguards (4 locales)
- data-centre towns de-fingerprinted; Hetzner named only in legal registers; fake K8s/HA
  infrastructure claims rewritten to the single-host truth; DR targets match the backup policy
- SPF canonicalised (amazonses.com); retention 730 everywhere; SLA unified; HIPAA/SIG claims
  aligned to "not offered"; /v1/messages endpoints; cancel route corrected; support SLAs unified
- Terms deduplicated; subprocessor changelog defabricated; SDK distribution honestly described
  (org repo is empty); footer build id, og:image SVG, Enterprise payment badge, case-studies
  route, AIism copy all fixed; og:image PNG everywhere; empty-social guards; site rebuilt

Console UX: demo KPIs/rows removed (real forms kept, honest empty state), quick-action forms
carry real inputs matching handler semantics, destructive Delete is a button-form, recovery
codes render as reveal-once chips, emoji API icons → labels

Hygiene: stale root /compliance fork, dead scrapers module + its tests, 1685 vendored dev
files removed; .env examples de-drifted

Full workspace tests green; pricing-drift, training-data, legal-identity and forbidden-pattern
validators pass.

+++ services/mail-server/crates/billing-service/src/types.rs
@@ -289,7 +289,7 @@ pub struct InvoiceLineItem {
     pub unit_price: i64,
     /// Line total in cents.
     pub amount: i64,
-    pub vat_rate: i32,
+    pub vat_rate: f64,
     pub vat_amount: i64,
 }
 
@@ -606,9 +606,9 @@ mod tests {
         let item = InvoiceLineItem {
             description: "Dedicated IP".into(),
             quantity: 2,
-            unit_price: 3000, // $30.00
+            unit_price: 3000, // €30.00
             amount: 6000,
-            vat_rate: 20,
+            vat_rate: 20.0,
             vat_amount: 1200,
         };
         assert_eq!(
